Corrected an oversight in the Factis spec
So, the way that I was testing the Factis module was to extend it
into an instance of a raw Object. While this worked and all, it
wasn't exactly what I was going for. This was meant to test that the
module could be properly extended into a class (traditional extend),
as opposed to an object (DCI-ish extend).

Granted, it's good to know that it works in either scenario, because
that means that it can be included via include for instances of a class
for things like Rspec examples.